add typing test
